% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sustainability_analyzer.R
\name{sustainability_analyzer}
\alias{sustainability_analyzer}
\title{sustainability_analyzer}
\description{
A class generator function
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Raw data storage}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sustainability_analyzer-new}{\code{sustainability_analyzer$new()}}
\item \href{#method-sustainability_analyzer-load_data}{\code{sustainability_analyzer$load_data()}}
\item \href{#method-sustainability_analyzer-validate_data}{\code{sustainability_analyzer$validate_data()}}
\item \href{#method-sustainability_analyzer-filter_data}{\code{sustainability_analyzer$filter_data()}}
\item \href{#method-sustainability_analyzer-calculate_kpis}{\code{sustainability_analyzer$calculate_kpis()}}
\item \href{#method-sustainability_analyzer-get_sites}{\code{sustainability_analyzer$get_sites()}}
\item \href{#method-sustainability_analyzer-get_date_range}{\code{sustainability_analyzer$get_date_range()}}
\item \href{#method-sustainability_analyzer-clone}{\code{sustainability_analyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-new"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-new}{}}}
\subsection{Method \code{new()}}{
Initialize the analyzer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-load_data"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-load_data}{}}}
\subsection{Method \code{load_data()}}{
Load and validate data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$load_data(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{Data frame with sustainability data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-validate_data"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-validate_data}{}}}
\subsection{Method \code{validate_data()}}{
Validate data structure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$validate_data(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{Data frame to validate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-filter_data"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-filter_data}{}}}
\subsection{Method \code{filter_data()}}{
Filter data by date range and sites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$filter_data(
  start_date = NULL,
  end_date = NULL,
  selected_sites = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{Start date for filtering}

\item{\code{end_date}}{End date for filtering}

\item{\code{selected_sites}}{Vector of selected sites}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-calculate_kpis"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-calculate_kpis}{}}}
\subsection{Method \code{calculate_kpis()}}{
Calculate KPIs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$calculate_kpis(filtered_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered_data}}{Filtered dataset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-get_sites"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-get_sites}{}}}
\subsection{Method \code{get_sites()}}{
Get unique sites
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$get_sites()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-get_date_range"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-get_date_range}{}}}
\subsection{Method \code{get_date_range()}}{
Get date range
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$get_date_range()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sustainability_analyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sustainability_analyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sustainability_analyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
